backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Data
yfinance==1.2.0
pandas==2.2.2
numpy==1.26.4
pandas-datareader==0.10.0

# ML
scikit-learn==1.4.2
lightgbm==4.3.0
statsmodels==0.14.2
scipy==1.13.0

# Scheduling
apscheduler==3.10.4

# HTTP / retries
httpx==0.27.0
tenacity==8.3.0

# AWS
boto3==1.34.84

# Config
python-dotenv==1.0.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
